+++ src/main/java/fr/estecka/variantscit/reload/LibraryDefinition.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Function;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.DataResult;
 import com.mojang.serialization.MapCodec;
@@ -16,17 +17,13 @@ import fr.estecka.variantscit.itemdata.transforms.IDataTransform;
 import fr.estecka.variantscit.itemdata.transforms.SuccessiveTransform;
 import fr.estecka.variantscit.itemdata.transforms.impl.StringCompareTransform;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.util.ExtraCodecs;
 
-/**
- * TODO: Fallback and special models are not included in this definition at this
- * time.
- */
+
 public record LibraryDefinition(
-	String modelPrefix,
+	Optional<String> modelPrefix,
 	IDataTransform namespacePredicate,
 	IDataTransform pathPredicate,
-	Map<ResourceLocation,ResourceLocation> hardcoded
+	Map<ResourceLocation,ResourceLocation> hardcodedList
 )
 {
 	static private final Codec<IDataTransform> TRANSFORM_CODEC = CodecUtil.WithAlternatives(
@@ -34,71 +31,67 @@ public record LibraryDefinition(
 		SuccessiveTransform.CODEC
 	);
 
-	static private final MapCodec<String> PREFIX_CODEC = CodecUtil.MapWithAlternative(
-		CodecUtil.IDENTIFIER_PATH.validate(CodecUtil.NonEmptyString("Model Prefix")).fieldOf("modelPrefix"),
-		Codec.BOOL.fieldOf("forceAllowEmptyPrefix").<String>flatXmap(
-			allowed -> allowed ? DataResult.success("") : DataResult.error(()->"Model Prefix cannot be empty."),
-			prefix -> DataResult.success(prefix.isEmpty())
+	static private final MapCodec<Optional<String>> PREFIX_CODEC = CodecUtil.MapWithAlternative(
+		CodecUtil.LEGACY_ITEM_PATH.optionalFieldOf("modelPrefix").validate(opt -> (opt.isEmpty() || !opt.get().isEmpty()) ? DataResult.success(opt) : DataResult.error(()->"Model Prefix cannot be empty")),
+		Codec.BOOL.fieldOf("forceAllowEmptyPrefix").flatXmap(
+			allowed -> allowed ? DataResult.success(Optional.of("")) : DataResult.error(()->"Model Prefix cannot be empty."),
+			prefix -> DataResult.success(prefix.isPresent() && prefix.get().isEmpty())
 		)
 	);
 
-	static public final MapCodec<LibraryDefinition> ADVANCED_MAPCODEC = RecordCodecBuilder.mapCodec(builder->builder
+	static private final Codec<Map<ResourceLocation,ResourceLocation>> LEGACY_SPECIAL_CODEC = Codec.unboundedMap(
+		CodecUtil.IDENTIFIER_PATH.flatXmap(path -> ResourceLocation.read("variants-cit:special/"+path), CodecUtil.NoGetter("Legacy Special")),
+		ResourceLocation.CODEC.validate(CodecUtil::UnItemify)
+	);
+
+	static private final Codec<Map<ResourceLocation,ResourceLocation>> MODELLIST_CODEC = CodecUtil.WithAlternatives(
+		Codec.unboundedMap(ResourceLocation.CODEC, ResourceLocation.CODEC),
+		ResourceLocation.CODEC.listOf().xmap(LibraryDefinition::FromArray, map->List.copyOf(map.keySet()))
+	);
+
+	static private final MapCodec<Map<ResourceLocation,ResourceLocation>> HARDCODED_CODEC = RecordCodecBuilder.<Map<ResourceLocation,ResourceLocation>>mapCodec(builder->builder
+		.group(
+			ResourceLocation.CODEC.validate(CodecUtil::UnItemify).optionalFieldOf("fallback").forGetter(CodecUtil.NoGetter("Legacy Fallback")),
+			LEGACY_SPECIAL_CODEC.optionalFieldOf("special", Map.of()).forGetter(CodecUtil.NoGetter("Legacy Special")),
+			MODELLIST_CODEC.optionalFieldOf("modelList", Map.of()).forGetter(Function.identity())
+		)
+		.apply(builder, LibraryDefinition::LegacyHardcoded)
+	);
+
+	static public final MapCodec<LibraryDefinition> MAP_CODEC = RecordCodecBuilder.mapCodec(builder->builder
 		.group(
 			PREFIX_CODEC.forGetter(LibraryDefinition::modelPrefix),
 			TRANSFORM_CODEC.optionalFieldOf("namespace", IDataTransform.NOOP).forGetter(LibraryDefinition::namespacePredicate),
 			IDataTransform.CODEC.optionalFieldOf("variantPathes", IDataTransform.NOOP).forGetter(LibraryDefinition::pathPredicate),
-			ExtraCodecs.strictUnboundedMap(ResourceLocation.CODEC, ResourceLocation.CODEC).fieldOf("hardcoded").forGetter(LibraryDefinition::hardcoded)
+			HARDCODED_CODEC.forGetter(LibraryDefinition::hardcodedList)
 		)
 		.apply(builder, LibraryDefinition::new)
 	);
 
-	static public final Codec<LibraryDefinition> ARRAY_CODEC = ResourceLocation.CODEC
-		.listOf()
-		.xmap(LibraryDefinition::FromArray, CodecUtil.NoGetter("Library Array form"))
-		;
-
-	static public final MapCodec<LibraryDefinition> ROOT_MAPCODEC = RecordCodecBuilder.mapCodec(builder->builder
-		.group(
-			CodecUtil.LEGACY_ITEM_PATH.validate(CodecUtil.NonEmptyString("Model Prefix")).fieldOf("modelPrefix").forGetter(LibraryDefinition::modelPrefix),
-			ResourceLocation.CODEC.validate(CodecUtil::UnItemify).optionalFieldOf("fallback").forGetter(CodecUtil.NoGetter("Legacy Fallback")),
-			Codec.unboundedMap(
-				CodecUtil.IDENTIFIER_PATH.flatXmap(path -> ResourceLocation.read("variants-cit:special/"+path), CodecUtil.NoGetter("Legacy Special")),
-				ResourceLocation.CODEC.validate(CodecUtil::UnItemify)
-			).optionalFieldOf("special", Map.of()).forGetter(CodecUtil.NoGetter("Legacy Special"))
-		)
-		.apply(builder, LibraryDefinition::FromLegacy)
-	);
-
-	static public final MapCodec<LibraryDefinition> MAP_CODEC = CodecUtil.MapWithAlternatives(
-		ADVANCED_MAPCODEC.fieldOf("assetLibrary"),
-		ARRAY_CODEC.fieldOf("assetLibrary"),
-		ROOT_MAPCODEC
-	);
-
 
 /******************************************************************************/
-/* Constructors                                                               */
+/* ModelList Constructors                                                     */
 /******************************************************************************/
 
-	static private LibraryDefinition FromArray(List<ResourceLocation> list){
+	static private Map<ResourceLocation,ResourceLocation> FromArray(List<ResourceLocation> list){
 		Map<ResourceLocation,ResourceLocation> hardcoded = new HashMap<>();
 		for (ResourceLocation id : list)
 			hardcoded.put(id, id);
-		return new LibraryDefinition("__noprefix__", IDataTransform.NULL, IDataTransform.NULL, hardcoded);
+		return hardcoded;
 	}
 
-	static private LibraryDefinition FromLegacy(String modelPrefix, Optional<ResourceLocation> fallback, Map<ResourceLocation,ResourceLocation> special){
-		Map<ResourceLocation,ResourceLocation> hardcoded = new HashMap<>(special);
+	static private Map<ResourceLocation,ResourceLocation> LegacyHardcoded(
+		Optional<ResourceLocation> fallback,
+		Map<ResourceLocation,ResourceLocation> special,
+		Map<ResourceLocation,ResourceLocation> hardcoded
+	){
+		hardcoded = new HashMap<>(hardcoded);
+		hardcoded.putAll(special);
 
 		if (fallback.isPresent())
 			hardcoded.put(VariantsCitMod.Identifier("fallback"), fallback.get());
 
-		return new LibraryDefinition(
-			modelPrefix,
-			IDataTransform.NOOP,
-			IDataTransform.NOOP,
-			Map.copyOf(hardcoded)
-		);
+		return Map.copyOf(hardcoded);
 	}
 
 
@@ -112,13 +105,13 @@ public record LibraryDefinition(
 	 */
 	public Set<ResourceLocation> GetVariantIds(ResourceLocation assetId){
 		Set<ResourceLocation> result = new HashSet<>();
-		if (assetId.getPath().startsWith(modelPrefix)){
-			ResourceLocation variantId = assetId.withPath(path->path.substring(modelPrefix.length()));
-			if (!this.hardcoded.containsKey(variantId) && this.AcceptsVariant(variantId))
+		if (modelPrefix.isPresent() && assetId.getPath().startsWith(modelPrefix.get())){
+			ResourceLocation variantId = assetId.withPath(path->path.substring(modelPrefix.get().length()));
+			if (!this.hardcodedList.containsKey(variantId) && this.AcceptsVariant(variantId))
 				result.add(variantId);
 		}
 
-		for (var entry : hardcoded.entrySet()) {
+		for (var entry : hardcodedList.entrySet()) {
 			if (entry.getValue().equals(assetId))
 				result.add(entry.getKey());
 		}